+++ lib/defdo/ddns/application.ex
@@ -6,6 +6,8 @@ defmodule Defdo.DDNS.Application do
   use Application
   require Logger
 
+  alias Defdo.DDNS.API.AuthConfig
+
   @impl true
   def start(_type, _args) do
     children =
@@ -30,10 +32,12 @@ defmodule Defdo.DDNS.Application do
 
   defp maybe_add_api_server(children) do
     if api_enabled?() do
+      validate_api_config!()
       Logger.info("Starting Defdo.DDNS API on port #{api_port()}")
 
       children ++
         [
+          Defdo.DDNS.API.AuthStore,
           {Bandit, plug: Defdo.DDNS.API.Router, scheme: :http, ip: {0, 0, 0, 0}, port: api_port()}
         ]
     else
@@ -41,9 +45,36 @@ defmodule Defdo.DDNS.Application do
     end
   end
 
+  @doc false
+  def validate_api_config!(config \\ nil) do
+    api_config = config || current_api_config()
+    enabled? = Keyword.get(api_config, :enabled, false)
+    allow_runtime_clients? = Keyword.get(api_config, :allow_runtime_clients, false)
+
+    case AuthConfig.from_config(api_config) do
+      {:ok, %{mode: _mode}} when not enabled? ->
+        :ok
+
+      {:ok, %{mode: mode}} when mode in [:token, :clients] ->
+        :ok
+
+      {:ok, %{mode: :none}} when allow_runtime_clients? ->
+        :ok
+
+      {:ok, %{mode: :none}} ->
+        raise ArgumentError,
+              "DDNS_API_TOKEN or DDNS_API_CLIENTS_JSON must be configured when DDNS_API_ENABLED=true"
+
+      {:error, {:invalid_clients, reason}} ->
+        raise ArgumentError, "DDNS_API_CLIENTS_JSON invalid: #{reason}"
+
+      {:error, reason} ->
+        raise ArgumentError, "DDNS API auth config invalid: #{inspect(reason)}"
+    end
+  end
+
   defp api_enabled? do
-    :defdo_ddns
-    |> Application.get_env(Defdo.DDNS.API, [])
+    current_api_config()
     |> Keyword.get(:enabled, false)
   end
 
@@ -56,8 +87,11 @@ defmodule Defdo.DDNS.Application do
   end
 
   defp api_port do
-    :defdo_ddns
-    |> Application.get_env(Defdo.DDNS.API, [])
+    current_api_config()
     |> Keyword.get(:port, 4050)
+  end
+
+  defp current_api_config do
+    Application.get_env(:defdo_ddns, Defdo.DDNS.API, [])
   end
 end